refactor server APIs to typescript
