+++ Server/API/balance.ts
@@ -1,7 +1,7 @@
 import fetch from 'node-fetch';
-import { DOGE_NETWORK } from './root.js';
+import { DOGE_NETWORK } from './root';
 
-export async function getBalance({ pubAddress }) {
+export async function getBalance(pubAddress: string) {
   const res = await fetch(
     `https://sochain.com/api/v2/get_address_balance/${DOGE_NETWORK}/${pubAddress}`
   );
@@ -9,7 +9,7 @@ export async function getBalance({ pubAddress }) {
   return json;
 }
 
-export async function getUnspentTx({ pubAddress }) {
+export async function getUnspentTx(pubAddress: string) {
   const res = await fetch(
     `https://sochain.com/api/v2/get_tx_unspent/${DOGE_NETWORK}/${pubAddress}`
   );
@@ -17,7 +17,13 @@ export async function getUnspentTx({ pubAddress }) {
   return json;
 }
 
-export async function isTxOutputSpent({ txId, outputNo }) {
+export async function isTxOutputSpent({
+  txId,
+  outputNo,
+}: {
+  txId: string;
+  outputNo: number;
+}) {
   const res = await fetch(
     `https://sochain.com/api/v2/is_tx_spent/${DOGE_NETWORK}/${txId}/${outputNo}`
   );
